Mocha: Allow tests to continue to run after a fail
In Mocha an error results in an exception getting thrown. Since this
exception was never caught there would always be the error item in the
queue that would prevent the next test from completing. The fix is to
make sure that we mark the test as done so that the next queued item can
be ran.