Upgrading to latest Steal
This updates FuncUnit to be based on the latest Steal and can now be npm
installed.

```js
var F = require("funcunit");
```

Bower still works as well.

Also added builds for AMD and CommonJS.

Tests are now modular and use `steal-qunit` and `steal-mocha`
respectively, config is based on the `package.json` although nothing
special is needed.